requirements.txt
pydantic
fastapi
finsim>=1.0.5
pandas
numpy
lppl
python-dotenv
